Aprender/Dica de projeto/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
playwright==1.44.0
python-docx==1.1.2
openpyxl==3.1.4
python-pptx==0.6.23
google-generativeai==0.8.3
openai==1.35.0
requests==2.32.3
python-dotenv==1.0.1
aiofiles==23.2.1
